Renamed skulptFileIO to cloudFileIO, and removed various old Skulpt-related code.
